refactor for simplification
